+++ packages/agent_core/__init__.py
@@ -1,4 +1,13 @@
-from .models import AgentResponse, HRDomain, HRRequest, RetrievedSource, RoutingResult
+from .models import (
+    AgentResponse,
+    EvaluationCase,
+    EvaluationResult,
+    EvaluationSummary,
+    HRDomain,
+    HRRequest,
+    RetrievedSource,
+    RoutingResult,
+)
 from .providers import MockProvider, ModelProvider, ProviderResponse
 from .retrieval import LocalKnowledgeRetriever
 from .router import ControllerRouter
@@ -6,6 +15,9 @@ from .workflow import HRAgentWorkflow
 
 __all__ = [
     "AgentResponse",
+    "EvaluationCase",
+    "EvaluationResult",
+    "EvaluationSummary",
     "ControllerRouter",
     "HRAgentWorkflow",
     "HRDomain",